requirement.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.1
discord-py-interactions==5.10.0
discord-typings==0.7.0
emoji==2.8.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.8.0
yarl==1.9.2
